move CompressedPostingList into a separate file
