+++ lib/segment/src/index/field_index/full_text_index/inverted_index.rs
@@ -3,11 +3,12 @@ use std::collections::{BTreeSet, HashMap};
 use common::types::PointOffsetType;
 use serde::{Deserialize, Serialize};
 
-use super::posting_list::{CompressedPostingList, PostingList};
+use super::posting_list::PostingList;
 use super::postings_iterator::{
     intersect_compressed_postings_iterator, intersect_postings_iterator,
 };
 use crate::common::operation_error::{OperationError, OperationResult};
+use crate::index::field_index::full_text_index::compressed_posting_list::CompressedPostingList;
 use crate::index::field_index::{CardinalityEstimation, PayloadBlockCondition, PrimaryCondition};
 use crate::types::{FieldCondition, Match, PayloadKeyType};
