+++ lib/segment/src/index/field_index/full_text_index/mod.rs
@@ -1,3 +1,4 @@
+mod compressed_posting_list;
 mod inverted_index;
 mod posting_list;
 mod postings_iterator;
